All comilation warnings were eliminated.

+++ src/hashset.h
@@ -295,7 +295,7 @@ const unsigned int  prime[24] = {
 	if(limit < prime[0])
 		return prime[0];
 
-	for(int i = 0; i < sizeof(prime) / sizeof(unsigned int); i++)
+    for(size_t i = 0; i < sizeof(prime) / sizeof(unsigned int); i++)
 	{
 		if(prime[i] > limit)
 			return prime[i];
